+++ .gitignore
@@ -19,6 +19,7 @@ out/
 # Gradle files
 .gradle/
 build/
+app/fabric.properties
 
 # Local configuration file (sdk path, etc)
 local.properties
