+++ app/build.gradle
@@ -10,7 +10,7 @@ buildscript {
 }
 apply plugin: 'com.android.application'
 apply plugin: 'io.fabric'
-
+Properties props = new Properties()
 android {
     compileSdkVersion 24
     buildToolsVersion "24.0.2"
@@ -24,18 +24,28 @@ android {
         targetSdkVersion 24
         versionCode 21
         versionName "1.0"
-
     }
     buildTypes {
         debug {
+            buildConfigField "String", "CONSUMER_KEY", "\"${props.getProperty("twitterConsumerKey")}\""
+            buildConfigField "String", "CONSUMER_SECRET", "\"${props.getProperty("twitterConsumerSecret")}\""
             ext.enableCrashlytics = false
         }
         release {
             minifyEnabled false
+            buildConfigField "String", "CONSUMER_KEY", "\"${props.getProperty("twitterConsumerKey")}\""
+            buildConfigField "String", "CONSUMER_SECRET", "\"${props.getProperty("twitterConsumerSecret")}\""
             proguardFiles getDefaultProguardFile('proguard-android.txt'), 'proguard-rules.pro'
         }
     }
 }
+
+try {
+    props.load(file('fabric.properties').newDataInputStream())
+} catch (Exception ex) {
+    throw new GradleException("Missing fabric.properties file in app");
+}
+
 repositories {
     jcenter()
     mavenCentral()
@@ -50,4 +60,5 @@ dependencies {
     }
     compile 'com.android.support:appcompat-v7:24.2.0'
     compile 'com.google.android.gms:play-services:9.4.0'
+    compile 'com.android.support:design:24.2.0'
 }
