+++ app/fabric.properties
@@ -2,3 +2,6 @@
 #Sun Sep 25 14:05:03 EDT 2016
 apiSecret=9a5649e1f9812dca63a6d9a3f90888009448c227de0f32741b4d7af591dce329
 twitterPluginId=57e8119d715ca57e0d00018e
+
+twitterConsumerKey=SArN2YtMrnaCgenNghXb0NKRC
+twitterConsumerSecret=S45LXQjBFbzb7PFC1UOvmeUs4WTzMjOpjnsMyIb3FG13oGqXO0
