+++ app/src/main/AndroidManifest.xml
@@ -10,6 +10,11 @@
     <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
     <uses-permission android:name="android.permission.INTERNET" />
 
+    <!-- To auto-complete the email text field in the login form with the user's emails -->
+    <uses-permission android:name="android.permission.GET_ACCOUNTS" />
+    <uses-permission android:name="android.permission.READ_PROFILE" />
+    <uses-permission android:name="android.permission.READ_CONTACTS" />
+
     <application
         android:allowBackup="true"
         android:icon="@mipmap/ic_launcher"
@@ -33,7 +38,7 @@
             android:value="bb159ef850c165e32df73a4534933fd777171a20" />
 
         <service
-            android:name=".TwitterService"
+            android:name=".FetchService"
             android:exported="false" />
 
         <activity
